feat(inventory-count): adjustments & stock checks - pending checks, reopen, box confirm (domain-gap 8, 10)

+++ backend/src/main/java/org/dawn/backend/controller/inventory/request/CreateStockAdjustmentRequest.java
@@ -1,10 +1,12 @@
 package org.dawn.backend.controller.inventory.request;
 
+import java.math.BigDecimal;
+
 public record CreateStockAdjustmentRequest(
         String type,
         Long productUnitId,
         Long productId,
-        Integer quantity,
+        BigDecimal quantity,
         String reason,
         String imageUrl,
         String sourceType,
