+++ backend/src/main/java/org/dawn/backend/controller/inventory/response/StockAdjustmentResponse.java
@@ -2,6 +2,7 @@ package org.dawn.backend.controller.inventory.response;
 
 import lombok.Builder;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 @Builder
@@ -14,7 +15,7 @@ public record StockAdjustmentResponse(
         Long productId,
         String productName,
         String productSku,
-        Integer quantity,
+        BigDecimal quantity,
         String reason,
         String imageUrl,
         Long locationId,
